requirements.txt
openai==0.28.0
pandas==1.5.3
transformers
sentence_transformers
pinecone-client
psycopg2-binary
streamlit
streamlit-option-menu
watchdog